move helper.EnsureWork() to pkg/controllers/binding
